split image.py functions into image_transform and image
